Update Lab.md
Because the method HasData() checks only on tokencache and on the login fase of the lab there is no tokencache yet. So you must check on cachedUserId = $"{userId}_UserCache"; but in code he will only check on this: cacheId = $"{userId}_TokenCache";. 

The result is that the solution or the lab not work and you will always stay on the Sign-In buttons.